% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H.R
\name{compute_H_W}
\alias{compute_H_W}
\title{Compute all the w_i blocks of H}
\usage{
compute_H_W(X, WHat, muHat, sigSqHat, K)
}
\arguments{
\item{X}{Data}

\item{WHat}{Loadings matrix}

\item{K}{Prior covariance matrix}
}
\value{
H_{w_i}
}
\description{
Compute all the w_i blocks of H
}
\examples{
set.seed(1)
d=10; k=3; n=10
X = matrix(rnorm(n*d), ncol=d)
W = matrix(rnorm(d*k), ncol=k)
mu = rnorm(d)
sigSq = rnorm(1)^2
K = cov.SE(matrix(1:10, ncol=1), beta=log(c(1, 3)))

library(numDeriv)
library(Matrix)
Hw1.analytic = stpca:::compute_H_W(X, W, mu, sigSq, K)[[1]]
Hw1.numeric  = Matrix(numDeriv::hessian(function(w) {
  W_ = W
  W_[,1] = w
  -(log_likelihood(X, W_, mu, sigSq) + log_prior(K, W_))
}, x=W[,1]))

stopifnot(all.equal(Hw1.analytic, Hw1.numeric))
}
