% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.exp.d}
\alias{cov.exp.d}
\title{Squared exponential covariance function partial derivatives wrt hyperparameters}
\usage{
cov.exp.d(X, X2, beta, D = NA, ...)
}
\description{
Squared exponential covariance function partial derivatives wrt hyperparameters
}
