% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.independent.d}
\alias{cov.independent.d}
\title{Derivative of the independent covariance function. Does not depend
on the value of \eqn{\sigma^2_k}; is always 1 everywhere the inputs have
zero distance, and zero everywhere else.}
\usage{
cov.independent.d(X, X2, beta, D = NA, ...)
}
\arguments{
\item{X}{Matrix of data}

\item{X2}{(optional) second matrix of data; if omitted, X is used.}

\item{beta}{The single hyperparameter: the log of the signal variance}
}
\description{
Derivative of the independent covariance function. Does not depend
on the value of \eqn{\sigma^2_k}; is always 1 everywhere the inputs have
zero distance, and zero everywhere else.
}
