% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.independent}
\alias{cov.independent}
\title{Independant covariance function. Is zero everywhere except for inputs with
zero distance. Has single hyperparameter of log signal variance. All nonzero
outputs are equal to the signal variance.}
\usage{
cov.independent(X, X2, beta, D = NA, ...)
}
\arguments{
\item{X}{Matrix of data}

\item{X2}{(optional) second matrix of data; if omitted, X is used.}

\item{beta}{The single hyperparameter: the log of the signal variance}
}
\description{
Independant covariance function. Is zero everywhere except for inputs with
zero distance. Has single hyperparameter of log signal variance. All nonzero
outputs are equal to the signal variance.
}
