% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.MR.beta0}
\alias{cov.MR.beta0}
\title{Computationally cheap estimate for beta0 for cov.MR}
\usage{
cov.MR.beta0(X, locations, k)
}
\arguments{
\item{X}{The dataset being analysed with stpca}

\item{locations}{Matrix containing the location of each feature in rows}

\item{k}{Latent dimensionality used in stpca}
}
\description{
Computationally cheap estimate for beta0 for cov.MR
}
