% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.MR.d}
\alias{cov.MR.d}
\title{Partial derivatives of the MR}
\usage{
cov.MR.d(X, beta)
}
\arguments{
\item{X}{Matrix of data}

\item{beta}{Hyperparameters; beta[1] is the log signal variance,
beta[-1] are the ncol(X) length scales.}
}
\description{
Partial derivatives of the MR
}
