% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.noisy.MR}
\alias{cov.noisy.MR}
\title{Noisy MR covariance function}
\usage{
cov.noisy.MR(X, beta)
}
\arguments{
\item{X}{Matrix of data}

\item{beta}{Hyperparameters; beta[1] is the log signal variance,
beta[2:(ncol(X)+1)] are the ncol(X) length scales. beta[ncol(X)+2]
is the noise variance}
}
\description{
Noisy MR covariance function
}
