% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.noisy.RQ.d}
\alias{cov.noisy.RQ.d}
\title{Partial derivatives of the noisy RQ covariance function}
\usage{
cov.noisy.RQ.d(X, X2, beta, D = NA, ...)
}
\arguments{
\item{X}{Matrix of data}

\item{X2}{(optional) second matrix of data; if omitted, X is used.}

\item{beta}{Hyperparameters; beta[1] is the log signal variance, beta[2] is
the log length scale, beta[3] is log alpha, and beta[4] is the
variance of the noise.}
}
\description{
Partial derivatives of the noisy RQ covariance function
}
