% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.noisy.RQ}
\alias{cov.noisy.RQ}
\title{Noisy RQ covariance function}
\usage{
cov.noisy.RQ(X, X2, beta, D = NA, ...)
}
\arguments{
\item{X}{Matrix of data}

\item{X2}{(optional) second matrix of data; if omitted, X is used.}

\item{beta}{Hyperparameters; beta[1] is the log signal variance, beta[2] is
the log length scale, beta[3] is log alpha, and beta[4] is the
variance of the noise.}
}
\description{
The sum of the Rational Quadratic and the independent covariance functions.
Interpreted as the smooth RQ covariance function plus iid noise. This is
implemented as the sum of two covariance functions, with the first three
elements of beta being sent to the RQ covariance function, and the last
element being sent to the independent covariance function.
}
