% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.RQ.beta0}
\alias{cov.RQ.beta0}
\title{Computationally cheap estimate for beta0 for cov.RQ.}
\usage{
cov.RQ.beta0(X, locations)
}
\description{
Computationally cheap estimate for beta0 for cov.RQ.
}
