% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.RQ}
\alias{cov.RQ}
\title{Rational quadratic covariance function}
\usage{
cov.RQ(X, X2, beta, D = NA, ...)
}
\description{
Rational quadratic covariance function
}
\examples{
locations = matrix(rnorm(10), ncol=2)
beta      = rnorm(3) # Logarithms of variance, length scale & alpha

K = cov.RQ(locations, beta=beta)
stopifnot(all(Matrix::diag(K)==exp(beta[1]))) # Diagonal is exp(beta[1])
stopifnot(all(svd(K)$d>0)) # K is positive definite
stopifnot(all(K[upper.tri(K)]<K[1,1])) # Largest element is on diagonal
}
