% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.SE.beta0}
\alias{cov.SE.beta0}
\title{Computationally cheap estimate for beta0 for cov.SE.}
\usage{
cov.SE.beta0(X, locations)
}
\description{
Computationally cheap estimate for beta0 for cov.SE.
}
