% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.SE.d}
\alias{cov.SE.d}
\title{Squared exponential covariance function derivatives wrt hyperparameters}
\usage{
cov.SE.d(X, X2, beta, D = NA, ...)
}
\description{
Squared exponential covariance function derivatives wrt hyperparameters
}
