% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.taper.d}
\alias{cov.taper.d}
\title{Taper a covariance function (partial derivatives)}
\usage{
cov.taper.d(cov.d, supLen = 1)
}
\arguments{
\item{cov.d}{The partial derivative function of the covariance function to
be tapered}

\item{supLen}{The support length of the MR}
}
\description{
Works as cov.taper, except this function takes a function producing the
partial derivatives of a covariance function. This modifies the function
to produce the partial derivatives of the tapered covariance function.
}
