% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.taper}
\alias{cov.taper}
\title{Taper a covariance function}
\usage{
cov.taper(cov, supLen = 1)
}
\arguments{
\item{cov}{The covariance function to be tapered}

\item{supLen}{The support length of the MR}
}
\description{
Takes a covariance function and tapers it by returning a new covariance
function which is the original multiplied by the MR covariance function. The
support length of the new covariance function will thus be the support
length of the MR (assuming that the original is not compactly supported).
}
