% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeightS.R
\name{createWeightS}
\alias{createWeightS}
\title{Create an asymmetric matrix of spatial weights from an \code{\link[SSN]{SSN}}
object.}
\usage{
createWeightS(ssndata, afvcol)
}
\arguments{
\item{ssndata}{filepath for an object of
\code{\link[SSN]{SpatialStreamNetwork-class}}.  Include the .ssn folder in
the path.}

\item{afvcol}{character string. This should be the column name in the SSN
object containing additive function values.}
}
\value{
Produces an asymmetric \eqn{p \times p} matrix (p = number of
  monitoring sites) where columns = upstream sites, rows = downstream sites.
}
\description{
The \code{createWeightS} function extracts the stream distance matrix for
observed data from a \code{\link[SSN]{SpatialStreamNetwork-class}} object and
constructs an asymmetric weight matrix based on the specified additive
function column.
}
\details{
Peterson and ver Hoef (2010) Appendix A shows how weights can be calculated
for the tail-up model, where weights include information about the flow
connected structure in the river network, network distance and relative
influence of upstream sites on downstream sites.  Weights are based on the
variable used to calculate the additive function in the SSN object.
\code{createWeightS} follows the steps described in Appendix A but stops
before the matrix is forced to symmetry.
}
\examples{
library(stpca)

## get filepath for SSN object
ssndata <- system.file("demoSSN/demoNet.ssn", package = "stpca")

## create matrix of spatial weights
weightS <- createWeightS(ssndata = ssndata, afvcol = "addfunccol")

}
\references{
Peterson, E. E. and J. M. ver Hoef (2010). A mixed-model moving
  average approach to geostatistical modeling in stream networks. Ecology 91,
  644-651.
}

