% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createWeightT.R
\name{createWeightT}
\alias{createWeightT}
\title{Create a matrix of temporal weights.}
\usage{
createWeightT(n, rho, corr.form = "AR1")
}
\arguments{
\item{n}{integer.  This should specify the number of time points in the data.}

\item{rho}{scalar.  Rho must be a value between 0 and 1 reflecting the
strength of temporal autocorrelation.}

\item{corr.form}{character string.  At present only "AR1" is implemented for
autoregressive temporal correlation of lag 1.}
}
\value{
An \eqn{n \times n} matrix of temporal weights.
}
\description{
The \code{createWeightT} function creates a symmetric matrix of temporal weights
based on temporal autocorrelation.   At present only AR(1) structure has been
implemented.
}
\details{
Description of AR1 correlation.
}
\examples{
weightT <- createWeightT(n = 5, rho = 0.5)
}

