% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoYmiss.R
\docType{data}
\name{demoYmiss}
\alias{demoYmiss}
\title{Spatiotemporal observations of a single water quality parameter.}
\format{A data.frame with 30 variables (columns, monitoring sites) and 25
  observations (rows, time points).  These time points are assumed to be
  regularly spaced so could represent weekly/monthly/annual intervals.
  \describe{ \item{x1 - x30}{25 regularly spaced temporal observations at
  sites x1 - x30} }}
\source{
See vignette for Creating Demo Data
}
\usage{
demoYmiss
}
\description{
A demo dataset of a single water quality parameter recorded at regularly
spaced time intervals and several monitoring sites, with 15\% of missing
values.
}
\keyword{datasets}

