% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-W.R
\name{dlaplace}
\alias{dlaplace}
\title{Density of the Laplace distribution}
\usage{
dlaplace(x, mu = 0, b = 1, logarithm = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{mu}{location, scalar}

\item{b}{scale, positive}

\item{logarithm}{logical; if TRUE, probabilities p are given as log(p)}
}
\description{
Returns the density Laplace(x | \\mu, b)
}
