% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_stpca.R
\name{fit_stpca}
\alias{fit_stpca}
\title{Update theta to be the maximum-a-posteriori value using Expectation
Maximisation.}
\usage{
fit_stpca(X, W, mu, sigSq, K, nIter = 50)
}
\arguments{
\item{nIter}{number of EM iterations}

\item{stpcaObj}{an stpca object.}
}
\value{
An \code{stpca} object.
}
\description{
Update theta to be the maximum-a-posteriori value using Expectation
Maximisation.
}
