% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppca-initialization.R
\name{initialize_from_ppca}
\alias{initialize_from_ppca}
\title{Initialize mu, sigSq and W from PPCA.}
\usage{
initialize_from_ppca(X, k)
}
\arguments{
\item{X}{Data}

\item{k}{latent dimensionality}
}
\value{
MAP PPCA parameters
}
\description{
Initialize mu, sigSq and W from PPCA.
}
