% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{log_prior_sigSq}
\alias{log_prior_sigSq}
\title{The improper prior over sigSq. Proportional to sigma^-2}
\usage{
log_prior_sigSq(sigSq)
}
\value{
un-normalised log prior (numeric)
}
\description{
The improper prior over sigSq. Proportional to sigma^-2
}
