% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{log_prior_W}
\alias{log_prior_W}
\title{The proper prior over W.
p(W) = \\prod^k_{i=1} N(w_i | 0, K)}
\usage{
log_prior_W(K, W)
}
\arguments{
\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}
}
\value{
normalised log prior (numeric)
}
\description{
The proper prior over W.
p(W) = \\prod^k_{i=1} N(w_i | 0, K)
}
