% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-W.R
\name{log_sparse_prior}
\alias{log_sparse_prior}
\title{The *un-normalised* prior over W, sigma^2 in SpStPCA. A product of
the proper, un-normalised p(W | \beta), and the improper p(sigma^2).}
\usage{
log_sparse_prior(K, W, sigSq, b)
}
\arguments{
\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}

\item{sigSq}{variance of noise added to 'signal' Wv_i}

\item{b}{Variance on Laplace part of prior}
}
\value{
un-normalised log prior over W
}
\description{
The *un-normalised* prior over W, sigma^2 in SpStPCA. A product of
the proper, un-normalised p(W | \beta), and the improper p(sigma^2).
}
