% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-W.R
\name{log_sparse_prior_W}
\alias{log_sparse_prior_W}
\title{The *un-normalised* prior over W in SpStPCA. This is a product of
the StPCA Gaussian prior and an iid Laplace prior.
p(W | beta, b) = \prod^k_{i=1} N(w_i | 0, K) *
                 \prod^{d}_{i=1} \prod^{k}_{j=1} Laplace(W_ij | 0, b)}
\usage{
log_sparse_prior_W(K, W, b)
}
\arguments{
\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}

\item{b}{Variance on Laplace part of prior}
}
\value{
un-normalised log prior over W
}
\description{
The *un-normalised* prior over W in SpStPCA. This is a product of
the StPCA Gaussian prior and an iid Laplace prior.
p(W | beta, b) = \prod^k_{i=1} N(w_i | 0, K) *
                 \prod^{d}_{i=1} \prod^{k}_{j=1} Laplace(W_ij | 0, b)
}
