% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{min.f.generator}
\alias{min.f.generator}
\title{Value and gradient of function to be minimized in tuning beta}
\usage{
\method{min}{f.generator}(X, W, mu, sigSq, locations, covar.fn, covar.fn.d,
  D = NA, max.dist = Inf, sparse = FALSE)
}
\value{
A function which, when given a value of beta, returns a value and
        gradient of a function to be minimized in beta-tuning.
}
\description{
Value and gradient of function to be minimized in tuning beta
}
\examples{
set.seed(1)
n = 100
d = 30
k = 3
X = matrix(rnorm(n*d), ncol=d)
W = matrix(rnorm(d*k), nrow=d, ncol=k)
mu = rnorm(d)
sigSq = rnorm(1)^2
locations = matrix(rnorm(d*2), ncol=2)
beta = rnorm(2)

fdf = stpca:::min.f.generator(X, W, mu, sigSq, locations, cov.SE, cov.SE.d)

library(numDeriv)
grad.analytic = fdf(beta)$df
grad.numeric = grad(function(beta_) {
  fdf(beta_)$f
}, x=beta)
all.equal(grad.analytic, grad.numeric, tolerance=100*.Machine$double.eps^0.5)
}
