% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sampNet.R
\name{plot_sampNet}
\alias{plot_sampNet}
\title{Plot the results from \code{\link{sampNet}}}
\usage{
plot_sampNet(samp.results, cov.params.plots = TRUE, pred.plots = TRUE,
  log.TUrange = FALSE, log.Eucrange = FALSE)
}
\arguments{
\item{samp.results}{character.  Contains the name of the object in which the
results from \code{\link{sampNet}} are stored.}

\item{cov.params.plots}{logical.  If \code{TRUE}, then plots of the
covariance parameters estimated for \code{nsim} sampled networks will be
produced.  See Details for further information.}

\item{pred.plots}{logical.  If \code{TRUE}, then plots of predicted values and
standard errors estimated for \code{nsim} sampled networks will be produced.  See
Details for further information.}

\item{log.TUrange}{logical.  If \code{TRUE}, then the y-axis and values
plotted in the Tail-up range parameter plot will be transformed using a
natural logarithmic transformation.  See Details.}

\item{log.Eucrange}{logical.  If \code{TRUE}, then the y-axis and values
plotted in the Euclidean range parameter plot will be transformed using a
natural logarithmic transformation.  See Details.}
}
\value{
Two sets of plots will be produced depending on argument values
  specified.
}
\description{
This function will plot the results from creating several sampled networks
using \code{\link{sampNet}}.  This will allow the user to make comparisons
between sampling schemes and proportion of monitoring sites retained in
sampled networks.
}
\details{
This function can be used to produce plots of several parameters/values
estimated for each \code{prop.sites} and \code{samp.scheme} combination.

If \code{cov.params.plots = TRUE} and \code{class(x)} is
\code{\link[SSN]{SpatialStreamNetwork-class}}, then five plots will be
produced showing the covariance function parameters estimated for \code{nsim}
sampled networks: Tail-up range, Euclidean range, Tail-up partial sill,
Euclidean partial sill, and nugget.  If \code{class(x)} is
\code{"data.frame"} then thre plots will be produced: range, partial sill,
and nugget.

If \code{pred.plots = TRUE}, then three plots will be produced: root mean
square predicted error (RMSPE) for observed locations, prediction error for
prediction locations, and average kriging standard error (AKSE) ratio.

RMSPE is calculated for observed locations as follows:

\deqn{\sqrt{\frac{1}{N}\sum^N_{i=1}\left(\hat{Y}(s_i)-Y(s_i)\right)^2}}

where \eqn{N} is the number of monitoring sites, \eqn{\hat{Y}(s_i)} is the
value at site \eqn{i} predicted from a model fitted to \eqn{\mathbf{s}_{-i}},
data from all monitoring sites except site \eqn{i}, and \eqn{Y(s_i)} is the
observed value at site \eqn{i}.  This is also known as Leave One Out Cross
Validation (LOOCV).  This is a measure of how well the model fits the data.
Low prediction error is desireable.

Prediction error is calculated for prediction locations as follows:

\deqn{\sqrt{\frac{1}{N_{pred}}\sum^{N_{pred}}_{i=1}\left(\hat{Y}_{full}(u_i)-\hat{Y}_{sampled}(u_i)\right)^2}}

where \eqn{N_{pred}} is the number of prediction locations,
\eqn{\hat{Y}_{full}(u_i)} is the value predicted at prediction (unobserved)
location \eqn{i} where the prediction is made using a model fitted to the
full monitoring network, and \eqn{\hat{Y}_{sampled}(u_i)} is the value
predicted at prediction (unobserved) location \eqn{i} where the prediction is
made using a model fitted to a sampled (reduced) network.  This is a measure
of how well the model based on fewer monitoring sites predicts at unobserved
locations compared to the model based on the full monitoring network.  As
with RMSPE, low values are desired.

AKSE ratio can be calculated as the ratio of AKSE from a model based on a
sampled (reduced) network to AKSE from a model based on the full monitoring
netowrk: \eqn{\mathrm{AKSE}_{sampled}/\mathrm{AKSE}_{full}} where

\deqn{\mathrm{AKSE} = \sqrt{\frac{1}{N_{pred}}\sum^{N_{pred}}_{i=1}\sigma^2(u_i)}}

and \eqn{N_{pred}} is the number of prediction locations, and
\eqn{\sigma^2(u_i)} is the kriging variance at prediction location \eqn{i}.

If \code{log.TUrange = TRUE} and \code{class(x)} is
\code{\link[SSN]{SpatialStreamNetwork-class}}, then values for the Tail-up
range parameter will have a natural log transformation applied.  This is
useful if the \code{nsim} estimated values vary by orders of magnitude and it
is left to the user to decide which display (transformed values or
untransformed values) is most useful for their own application.  The
transformation is also available for the Euclidean range parameter for
\code{class(x)} is \code{\link[SSN]{SpatialStreamNetwork-class}} or
\code{"data.frame"}.
}
\examples{
data(sampDemo)
plot_sampNet(sampDemo,
             cov.params.plots = TRUE,
             pred.plots = TRUE)

}

