% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stpca.R
\name{plot_stpca}
\alias{plot_stpca}
\title{Visualize (weighted) PCA results.}
\usage{
plot_stpca(x, plots = c("map", "biplot", "glyph", "ts", "meanPM"),
  pc.map = 1, probs = c(0, 0.25, 0.75, 1), pc.biplot = c(1, 2),
  biplot.factor = 100, pc.glyph = c(1, 2, 3), river, r1 = 10, coords,
  pc.ts = c(1, 2, 3), pc.meanPM = c(1, 2, 3), meanPM.factor = 0.05)
}
\arguments{
\item{x}{character string.  This should specify the name of the object in
which the results of \code{\link{stpca}} are stored.}

\item{plots}{character string.  Specifies the type of plot to be produced.
Can include any combination of c("map", "biplot", "glyph", "ts", "meanPM").  See
Details for further information.}

\item{pc.map}{vector (of length one).  Specifies which principal component to
plot on a map.  Default is PC1.}

\item{probs}{vector.  A vector of probabilities used to calculate color
breaks when \code{plots = "map"}.  Values should be in the range [0,1].}

\item{pc.biplot}{vector.  Specifies the principal components to plot when
\code{plots = "biplot"}.  Default is PC1 on x-axis and PC2 on y-axis.}

\item{biplot.factor}{scalar.  A scaling value used to produce the biplot.
Higher values make arrows shorter.}

\item{pc.glyph}{vector.  Specifies the principal components to plot when
\code{plots = "glyph"}.  It is recommended that the glyph plot is used only
to display 3 or more principal components.}

\item{river}{character string.  Name of object containing river network
shapefile.  Used for plotting results on a map of the river network where
\code{plots = c("map", "glyph")}.}

\item{r1}{scalar.  Used to control relative lengths of spokes on glyphs in
glyph plot.  Larger values decrease the size of the glyphs.  Default is 10.}

\item{coords}{matrix of 2 columns.  The columns should contain the x- and
y-axis coordinates for the monitoring site locations respectively and MUST
be in the same order as monitoring sites are specified in rows
(\code{pca.mode = "Tmode"}) or columns (\code{pca.mode = "Smode"}) in the
dataframe specified in \code{\link{stpca}}.}

\item{pc.ts}{vector.  Specifies which principal components should have
principal components/scores plotted as time series.  Should only be used
when \code{pca.mode = "Smode"} in \code{\link{stpca}}.  Default is
c(1,2,3).}

\item{pc.meanPM}{vector.  Specifies which principal components should be
plotted.  Should only be used when \code{pca.mode = "Smode"} in
\code{\link{stpca}}.  Default is c(1,2,3).}

\item{meanPM.factor}{scalar.  Used to control the contrast between the mean
time series over all sites and +/- a multiple of the principal
components/scores.  See Details.}
}
\value{
Produces the specified plot(s) for the principal components specified
  and for each weighting scheme used in \code{\link{stpca}}.
}
\description{
Visualize the results from applying PCA, adjusted for spatial and/or temporal
weights if appropriate.  This will plot the results from the output of
\code{\link{stpca}}.
}
\details{
The type of plot to use will depend on how many principal components are of
interest.  If only 1 PC is of interest then \code{plots = "map"} should be
used, or \code{plots = "biplot"} if a pair of PC's  should be plotted.  For 3
or more PC's then \code{plots = "glyph"} is recommended.

\code{"map"} will display loadings (S-mode PCA) or scores (T-mode PCA) for a
single PC.  Color breaks are specified using the \code{probs} argument.

\code{"biplot"} displays the standard biplot, commonly used to investigate
the results of PCA and displays both scores and loadings on the same plot,
for any 2 PC's specified using \code{pc.biplot}.  Arrows represent loadings
and points represent principal component scores.

\code{"glyph"} is based on \code{\link[GWmodel]{glyph.plot}} in the GWmodel
package, and displays a map of the river network (if \code{river} is
specified) and glyphs showing the loadings (S-mode PCA) or scores (T-mode
PCA) for the principal components specified using \code{pc.glyph}.  The
lengths of the spokes are relative to each other and the spoke at the 12
o'clock position represents the first PC specified in \code{pc.glyph}, and
other PC's are represented moving round clockwise.  Blue indicates positive
values and red indicates negative values.  A message will appear when glyph
plots are produced prompting the user to check that the matrix of coordinates
has monitoring sites in the same order as in the rows (\code{pca.mode =
"Tmode"}) or columns (\code{pca.mode = "Smode"}) of the dataframe used for
\code{\link{stpca}}.  Incorrect ordering will result in glyphs being
displayed at the wrong locations.  If a glyph plot is produced for fewer than
three PC's then the plot will be produced but a message suggesting other
plots might be more appropriate will also appear.  Glyph plots and code are
based on Isabella et. al. (2015).

\code{"ts"} displays a time series plot of the principal components when
\code{pca.mode = "Smode"} in \code{\link{stpca}}.  This plot is not produced
for \code{pca.mode = "Tmode"} since a time series of loadings has no
meaningful interpretation.  Rembember that since loadings can be positive or
negative, the temporal pattern described by each PC might the negative of
what is displayed on the plot.

\code{"meanPM"} displays the mean time series for all sites, +/- a multiple
(meanPM.factor) of the principal components specified using \code{pc.meanPM}.
Red symbols are the mean time series + (meanPM.factor*scores), blue symbols
are the mean time series - (meanPM.factor*scores).  This plot shows the
variation around the mean captured by each principal component.  The pattern
might indicate that a principal component describes a shift in the mean,
where some monitoring sites have generally higher/lower values than the mean.
The principal component might describe a dampening of the mean signal, where
the seasonal pattern oscillates at a lower frequency than the mean pattern.
Another possibility is that the prinicpal compnent captures a time lag at
some sites compared to the mean.  This list is not exhaustive but has been
included to provide some suggestions for interpretation.
}
\examples{
library(stpca)
library(maptools) #for reading in shapefile
library(SSN) #to import SSN object

## get data.frame with no missing values
data(demoY)

## unweighted T-mode PCA
Tmode.pca.uw <- stpca(x=demoY, pca.mode="Tmode", pca.wt=c("unweighted"))

## spatial, temporal, and spatiotemporal weighted S-mode PCA

# create spatial weights
ssndata <- system.file("demoSSN/demoNet.ssn", package="stpca")
weightS <- createWeightS(ssndata=ssndata, afvcol="addfunccol")

# create temporal weights
weightT <- createWeightT(n=nrow(demoY), rho=0.5)

# check that the order of monitoring sites in x is the same as the
# columns in spatial.wt.  They are the same if the following results
# in 1.
mean(colnames(demoY) == colnames(weightS))

# weighted S-mode PCA
Smode.pca.all <- stpca(x = demoY, pca.mode = "Smode",
                       spatial.wt = weightS, temporal.wt = weightT,
                       pca.wt = c("unweighted", "spatial", "temporal",
                                  "spatiotemporal"))

## get the river network shapefile, in this case the edges information in
## demoNet
demoNet.path <- system.file("demoSSN/demoNet.ssn", package="stpca")
river.path <- paste(demoNet.path, "edges.shp", sep="/")
river <- readShapeLines(river.path)
demoNet <- importSSN(demoNet.path)
data1 <- getSSNdata.frame(demoNet, "Obs")
coords <- c("NEAR_X", "NEAR_Y")
coords <- data1[,coords]

## plot the results

# unweighted T-mode PCA
plot_stpca(x = Tmode.pca.uw, plots = "map", coords = coords,
           river = river, pc.map = 1, probs = seq(0, 1, by=0.5))
plot_stpca(x = Tmode.pca.uw, plots = "biplot")
plot_stpca(x = Tmode.pca.uw, plots = "glyph", pc.glyph=1:2,
           river = river, coords = coords, r1=5)


# spatial, temporal, and spatiotemporal weighted S-mode PCA
plot_stpca(x = Smode.pca.all, plots = "map", coords = coords,
           river = river)
plot_stpca(x = Smode.pca.all, plots = "biplot")
plot_stpca(x = Smode.pca.all, plots = "glyph", river = river,
           coords = coords)
plot_stpca(x = Smode.pca.all, plots = "ts", pc.ts = 1:3)
plot_stpca(x = Smode.pca.all, plots = "meanPM")

}
\references{
Biplot code is based on:

\url{http://steviep42.bitbucket.org/YOUTUBE.DIR/BB_phys_stats_ex1.R}
and
\url{https://www.youtube.com/watch?v=I5GxNzKLIoU}.

Glyph plot based on:

Isabella Gollini, Binbin Lu, Martin Charlton, Christopher Brunsdon, Paul
Harris (2015).  GWmodel: An R Package for Exploring Spatial Heterogeneity
Using Geographically Weighted Models.   Journal of Statistical Software,
63(17), 1-50. \url{http://www.jstatsoft.org/v63/i17/}.
}

