% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{predict.spca}
\alias{predict.spca}
\title{De-noise a sample using a trained \code{spca} object.}
\usage{
\method{predict}{spca}(object, ...)
}
\arguments{
\item{object}{An \code{spca} object returned from a call to \code{spca}}

\item{samples}{samples to de-noise}
}
\value{
De-noised samples of the same dimensionality as the parameter \code{samples}
}
\description{
De-noise a sample using a trained \code{spca} object.
}
