% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.stpca}
\alias{predict.stpca}
\title{De-noise a sample using a trained \code{stpca} object.}
\usage{
\method{predict}{stpca}(object, ...)
}
\arguments{
\item{object}{An \code{stpca} object returned from a call to \code{stpca}}

\item{samples}{samples to de-noise}
}
\value{
De-noised samples of the same dimensionality as the parameter \code{samples}
}
\description{
De-noise a sample using a trained \code{stpca} object.
}
