% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse-W.R
\name{soft_threshold}
\alias{soft_threshold}
\title{Soft-thresholding operator.}
\usage{
soft_threshold(x, threshold)
}
\arguments{
\item{x}{the vector/matrix of values to be soft thresholded}

\item{threshold}{the threshold lambda}
}
\description{
The soft-thresholding operator S_\\lambda(x) = sign(x)(abs(x) - \\lambda)_+
}
