% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca.continue}
\alias{spca.continue}
\title{Does not yet work with 'new' SPCA architecture}
\usage{
spca.continue(spcaObj, trace = 0, report_iter = 10, max.dist = Inf,
  maxit = 10, tol = 0.01, ucminf.control = list())
}
\description{
Does not yet work with 'new' SPCA architecture
}
