% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{spca.H}
\alias{spca.H}
\title{Compute all the blocks of H.}
\usage{
spca.H(X, W, mu, sigSq, K)
}
\arguments{
\item{X}{Data}

\item{W}{Loadings matrix}

\item{K}{Prior covariance matrix}
}
\value{
H
}
\description{
Compute all the blocks of H.
}
\examples{
set.seed(1)
d=10; k=3; n=1000
X = matrix(rnorm(n*d), ncol=d)
W = matrix(rnorm(d*k), ncol=k)
mu = rnorm(d)
sigSq = rnorm(1)^2
K = cov.SE(matrix(1:10, ncol=1), beta=log(c(2, 3)))

library(numDeriv)
library(Matrix)

#Test that the analytic hessian for mu & sigSq matches numerical Hessian.
H.analytic = spca:::spca.H(X, W, mu, sigSq, K)
HsigSq.numeric = Matrix(hessian(function(sigSq_) {
  -spca.log_posterior(X, K, W, mu, sigSq_)
}, x=sigSq))
stopifnot(all.equal(H.analytic$sigSq, HsigSq.numeric,
                    tolerance=1e-8))

Hmu.numeric = Matrix(hessian(function(mu_) {
  -spca.log_posterior(X, K, W, mu_, sigSq)
}, x=mu))
stopifnot(all.equal(H.analytic$mu, Hmu.numeric, tolerance=1e-6))
}
