% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{spca.H.W}
\alias{spca.H.W}
\title{Internal function for computing the w_i block of H}
\usage{
spca.H.W(X, W, mu, sigSq, K)
}
\arguments{
\item{X}{Data}

\item{W}{Loadings matrix}

\item{K}{Prior covariance matrix}
}
\value{
H_{w_i}
}
\description{
Internal function for computing the w_i block of H
}
