% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca.log_bayes_factor}
\alias{spca.log_bayes_factor}
\title{Compute bayes factor}
\usage{
spca.log_bayes_factor(X, K1, W1, mu1, sigSq1, K2, W2, mu2, sigSq2)
}
\arguments{
\item{X}{Data}
}
\value{
Log bayes factor; model 1 is numerator
}
\description{
Compute bayes factor
}
