% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{spca.log_evidence}
\alias{spca.log_evidence}
\title{Compute the laplace approximation to the log evidence given the MAP
parameters K, mu, sigSq as well as the prior covariance matrix K.
Note that this is multiplied by an UN-KNOWN CONSTANT due to the flat
priors over mu and sigSq. However, this unknown constant is always
the same regardless of k and K, so this may be used to compute
meaningful bayes factors between SPCA models.}
\usage{
spca.log_evidence(X, K, W, mu, sigSq)
}
\arguments{
\item{X}{Data}

\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}
}
\value{
Approximate log evidence
}
\description{
Compute the laplace approximation to the log evidence given the MAP
parameters K, mu, sigSq as well as the prior covariance matrix K.
Note that this is multiplied by an UN-KNOWN CONSTANT due to the flat
priors over mu and sigSq. However, this unknown constant is always
the same regardless of k and K, so this may be used to compute
meaningful bayes factors between SPCA models.
}
