% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca.log_likelihood}
\alias{spca.log_likelihood}
\title{Calculate the log likelihood for SPCA with given parameters}
\usage{
spca.log_likelihood(X, W, mu, sigSq)
}
\arguments{
\item{X}{Data}
}
\value{
log likelihood (numeric)
}
\description{
Calculate the log likelihood for SPCA with given parameters
}
