% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca.log_posterior}
\alias{spca.log_posterior}
\title{Calculate the *un-normalised* log posterior for SPCA with given parameters}
\usage{
spca.log_posterior(X, K, W, mu, sigSq)
}
\arguments{
\item{X}{Data}

\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}
}
\value{
un-normalised log posterior (numeric)
}
\description{
Calculate the *un-normalised* log posterior for SPCA with given parameters
}
