% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca.log_prior}
\alias{spca.log_prior}
\title{Calculate the *un-normalised* log prior for SPCA with given loadings matrix}
\usage{
spca.log_prior(K, W)
}
\arguments{
\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}
}
\value{
un-normalised log prior (numeric)
}
\description{
Calculate the *un-normalised* log prior for SPCA with given loadings matrix
}
