% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{Performs SPCA}
\usage{
spca(X, k, locations, covar.fn, covar.fn.d = NULL, beta0 = c(), trace = 0,
  report_iter = 10, max.dist = Inf, maxit = 20, maxit.outer = 5)
}
\arguments{
\item{X}{Data}

\item{k}{Latent dimensionality}

\item{locations}{the coordinates of each dimenion in X}

\item{covar.fn}{covariance function to generate K_\beta}

\item{covar.fn.d}{gradient of the covariance function with respect to hyperparameters}

\item{beta0}{initial hyperparameters}

\item{trace}{amount of reporting. 0=none, 1=low, 2=high}

\item{report_iter}{Number of iterations between reports}

\item{max.dist}{Maximum distance between features to consider}

\item{maxit}{number of inner iterations}

\item{maxit.outer}{number of outer iterations}
}
\value{
An \code{spca} object.
}
\description{
Performs SPCA
}
