% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpca.init.R
\name{stpca.init}
\alias{stpca.init}
\title{Initialises an stpca object.}
\usage{
stpca.init(X, k, locations, covar.fn, covar.fn.d = NULL, beta0 = c(),
  trace = 0, max.dist = Inf, constraints = NULL)
}
\arguments{
\item{X}{Data}

\item{k}{Latent dimensionality}

\item{locations}{the coordinates of each dimenion in X}

\item{covar.fn}{covariance function to generate K_beta}

\item{covar.fn.d}{gradient of the covariance function with respect to hyperparameters}

\item{beta0}{initial hyperparameters}

\item{trace}{amount of reporting. 0=none, 1=low, 2=high}

\item{max.dist}{Maximum distance between features to consider}

\item{constraints}{constraints on beta, of the form used in maxLik}
}
\value{
An \code{stpca} object.
}
\description{
Initialises an stpca object.
}
