% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpca.iterate.R
\name{stpca.iterate.beta}
\alias{stpca.iterate.beta}
\title{Update beta to be the maximum marginal likelihood value using BFGS.}
\usage{
stpca.iterate.beta(stpcaObj)
}
\arguments{
\item{stpcaObj}{an stpca object.}
}
\value{
An \code{stpca} object.
}
\description{
Update beta to be the maximum marginal likelihood value using BFGS.
}
