% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpca.iterate.R
\name{stpca.iterate.theta}
\alias{stpca.iterate.theta}
\title{Update theta to be the maximum-a-posteriori value using Expectation
Maximisation.}
\usage{
stpca.iterate.theta(stpcaObj, maxit.inner = 10)
}
\arguments{
\item{stpcaObj}{an stpca object.}

\item{maxit.inner}{number of EM iterations}
}
\value{
An \code{stpca} object.
}
\description{
Update theta to be the maximum-a-posteriori value using Expectation
Maximisation.
}
