% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{stpca.log_det_H_d}
\alias{stpca.log_det_H_d}
\title{Compute the partial derivatives of log(det(H)) with respect to the
hyperparameters with the value of beta provided.}
\usage{
stpca.log_det_H_d(K, dK, HW)
}
\arguments{
\item{K}{Prior covariance matrix}

\item{dK}{Prior covariance matrix derivatives}

\item{HW}{list of blocks H_{w_i}}
}
\value{
Partial derivatives of log|H|
}
\description{
Compute the partial derivatives of log(det(H)) with respect to the
hyperparameters with the value of beta provided.
}
