% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{stpca.log_evidence_d}
\alias{stpca.log_evidence_d}
\title{Compute the derivative of the approximate log evidence with respect to the
value of beta provided.}
\usage{
stpca.log_evidence_d(X, K, W, mu, sigSq, beta, dK)
}
\arguments{
\item{X}{Data}

\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}
}
\value{
Partial derivatives of approximate log evidence
}
\description{
Compute the derivative of the approximate log evidence with respect to the
value of beta provided.
}
