% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-quantities.R
\name{stpca.log_likelihood}
\alias{stpca.log_likelihood}
\title{Calculate the log likelihood for StPCA with given parameters}
\usage{
stpca.log_likelihood(X, W, mu, sigSq)
}
\arguments{
\item{X}{Data}
}
\value{
log likelihood (numeric)
}
\description{
Calculate the log likelihood for StPCA with given parameters
}
\examples{
d = 50
k = 5
n = 15

set.seed(1)
X  = matrix(rnorm(n*d), nrow=n, ncol=d)
mu = colMeans(X)
Xc = sweep(X, 2, mu, '-')
covar.svd = svd(Xc/sqrt(n), nu=0, nv=k)
covar.eigval = covar.svd$d^2
sigSq = sum(covar.eigval[-(1:k)])/(d-k)
W     = covar.svd$v \%*\% diag(sqrt(covar.eigval[1:k] - sigSq), ncol=k, nrow=k)

R     = svd(matrix(rnorm(k*k), ncol=k, nrow=k))$u # Random orthonormal matrix

# The likelihood is invariant to multiplying  by an orthonormal matrix.
l1 = stpca.log_likelihood(X, W, mu, sigSq)
l2 = stpca.log_likelihood(X, W\%*\%R, mu, sigSq)
stopifnot(all.equal(l1, l2))
}
