% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical-quantities.R
\name{stpca.log_posterior}
\alias{stpca.log_posterior}
\title{Calculate the *un-normalised* log posterior for StPCA with given parameters}
\usage{
stpca.log_posterior(X, K, W, mu, sigSq)
}
\arguments{
\item{X}{Data}

\item{K}{Prior covariance matrix}

\item{W}{Loadings matrix}
}
\value{
un-normalised log posterior (numeric)
}
\description{
Calculate the *un-normalised* log posterior for StPCA with given parameters
}
