% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize-data.R
\name{synthesize_data}
\alias{synthesize_data}
\title{Synthesize fake data from SPCA model}
\usage{
synthesize_data(n, k, K, noisesd = 0)
}
\arguments{
\item{n}{Number of samples to create}

\item{k}{Latent dimensionality}

\item{K}{prior covariance matrix}

\item{noisesd}{standard deviation of noise added to data}
}
\description{
Synthesize fake data from SPCA model
}
