% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthesize-data.R
\name{synthesize_data_kern}
\alias{synthesize_data_kern}
\title{Synthesize fake data from SPCA model}
\usage{
synthesize_data_kern(n, k, dim, kern, noisesd = 0)
}
\arguments{
\item{n}{Number of samples to create}

\item{k}{Latent dimensionality}

\item{dim}{dimensions of grid to construct}

\item{kern}{cvarinace function used to build the prior covariance matrix K}

\item{noisesd}{standard deviation of noise added to data}
}
\description{
Synthesize fake data from SPCA model
}
