% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM.R
\name{theta_EM}
\alias{theta_EM}
\title{Update theta to be the maximum-a-posteriori value using Expectation
Maximisation.}
\usage{
theta_EM(X, W, mu, sigSq, K, maxit = 500, bftol = 1e-05)
}
\description{
Update theta to be the maximum-a-posteriori value using Expectation
Maximisation.
}
