% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.RQ.beta0}
\alias{cov.RQ.beta0}
\title{Computationally cheap estimate for beta0 for cov.RQ.}
\usage{
cov.RQ.beta0(X, locations, k)
}
\arguments{
\item{X}{The dataset being analysed with stpca}

\item{locations}{Matrix containing the location of each feature in rows}

\item{k}{Latent dimensionality used in stpca}
}
\description{
Computationally cheap estimate for beta0 for cov.RQ.
}
\examples{
# Construct some synthetic data, initialise beta for the RQ kernel from this dataset.
library(functional)
n = 10; k = 4; dim=c(10, 10)
kern = Curry(cov.noisy.SE, beta=log(c(2, 0.4, 0.1)))
synth = synthesize_data_kern(n, k, dim, kern, noisesd=0.2)
beta0 = cov.RQ.beta0(synth$X, synth$grid, k)
stopifnot(length(beta0) == 3)
stopifnot(all(is.finite(beta0)))
}
