% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.RQ.d}
\alias{cov.RQ.d}
\title{Rational quadratic covariance function derivatives wrt hyperparameters}
\usage{
cov.RQ.d(X, beta, D = NA, ...)
}
\description{
Rational quadratic covariance function derivatives wrt hyperparameters
}
