% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariance-functions.R
\name{cov.SE}
\alias{cov.SE}
\title{Squared exponential covariance function}
\usage{
cov.SE(X, X2, beta, D = NA, ...)
}
\arguments{
\item{X}{Matrix of data}

\item{X2}{(optional) second matrix of data; if omitted, X is used.}

\item{beta}{Hyperparameters; beta[1] is the log signal variance, beta[2] is the log length scale.}
}
\description{
Squared exponential covariance function
}
\examples{
# Confirm that diagonal of covariance matrix is the specified variance, and
# the off-diagonals are less than this.
grid = matrix(1:10, ncol=1)
beta = rnorm(2)
K    = cov.SE(grid, beta=beta)
stopifnot(all(Matrix::diag(K)==exp(beta[1])))
stopifnot(all(K[upper.tri(K)]<exp(beta[1])))
}
