% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-approximation.R
\name{spca.H.W}
\alias{spca.H.W}
\title{Compute all the w_i blocks of H}
\usage{
spca.H.W(X, W, mu, sigSq, K)
}
\arguments{
\item{X}{Data}

\item{W}{Loadings matrix}

\item{K}{Prior covariance matrix}
}
\value{
H_{w_i}
}
\description{
Compute all the w_i blocks of H
}
\examples{
set.seed(1)
d=10; k=3; n=10
X = matrix(rnorm(n*d), ncol=d)
W = matrix(rnorm(d*k), ncol=k)
mu = rnorm(d)
sigSq = rnorm(1)^2
K = cov.SE(matrix(1:10, ncol=1), beta=log(c(1, 3)))

library(numDeriv)
library(Matrix)
Hw1.analytic = spca:::spca.H.W(X, W, mu, sigSq, K)[[1]]
Hw1.numeric  = Matrix(hessian(function(w) {
  W_ = W
  W_[,1] = w
  -spca.log_posterior(X, K, W_, mu, sigSq)
}, x=W[,1]))

stopifnot(all.equal(Hw1.analytic, Hw1.numeric))
}
