% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spca.R
\name{spca}
\alias{spca}
\title{Performs SPCA}
\usage{
spca(X, k, locations, covar.fn, covar.fn.d = NULL, beta0 = c(), trace = 0,
  report_iter = 10, max.dist = Inf, maxit = 20, maxit.outer = 5)
}
\arguments{
\item{X}{Data}

\item{k}{Latent dimensionality}

\item{locations}{the coordinates of each dimenion in X}

\item{covar.fn}{covariance function to generate K_beta}

\item{covar.fn.d}{gradient of the covariance function with respect to hyperparameters}

\item{beta0}{initial hyperparameters}

\item{trace}{amount of reporting. 0=none, 1=low, 2=high}

\item{report_iter}{Number of iterations between reports}

\item{max.dist}{Maximum distance between features to consider}

\item{maxit}{number of inner iterations}

\item{maxit.outer}{number of outer iterations}
}
\value{
An \code{spca} object.
}
\description{
Performs SPCA
}
\examples{
library(fields)
data(ozone2) # TODO: Get rid of dependancy on fields package

# Missing data: Replace missing values by column means
X = ozone2$y
for (col in 1:ncol(X)) {
  ind = is.na(X[,col])
  X[ind,col] = mean(X[,col], na.rm=TRUE)
}
X = X/sd(X) # Scale for numerical reasons

locations = ozone2$lon.lat
locations = apply(locations, 2, function(col) (col-min(col))/(max(col)-min(col)))

model.spca = spca(X, 3, locations, cov.SE, cov.SE.d, beta0=log(c(1, 0.5)),
                  maxit=20, maxit.outer=3, trace=0)
}
