% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpca.R
\name{stpca}
\alias{stpca}
\title{Perform PCA to identify common spatiotemporal patterns, adjusting for spatial
and/or temporal autocorrelation where appropriate.}
\usage{
stpca(x, center = TRUE, scale = FALSE, pca.mode = NULL,
  pca.wt = c("unweighted"), spatial.wt, temporal.wt, scree = TRUE, k = 10)
}
\arguments{
\item{x}{data.frame.  This should be a complete data set with no missing
values.  Columns MUST be monitoring sites while rows MUST be time points.
See Details for further information.}

\item{center}{logical. TRUE (default) will subtract column means from each
column of the \code{x}.}

\item{scale}{logical.  TRUE will divide values in \code{x} by the column
standard deviation.  Default is FALSE.  See Details for further
information.}

\item{pca.mode}{character string.  MUST be specified.  Options are "Smode" or
"T-mode".  See Details for further information.}

\item{pca.wt}{character vector.  "unweighted" (default) performs standard
PCA, "spatial" adjusts for spatial autocorrelation, "temporal" adjusts for
temporal autocorrelation, "spatiotemporal" adjusts for both spatial and
temporal autocorrelation.  See Details for further information.}

\item{spatial.wt}{character string. Specifies the name of the matrix
containing spatial weights.  Only used if \code{pca.wt} contains
"spatial" or "spatiotemporal".   Can be created by the user or using
\code{\link{createWeightS}}.}

\item{temporal.wt}{character string. Specifies the name of the matrix
containing temporal weights.  Only used if \code{pca.wt} contains
"temporal" or "spatiotemporal".  Can be created by the user or using
\code{\link{createWeightT}}.}

\item{scree}{logical. TRUE (default) means a scree plot will be produced for
the first k principal components.}

\item{k}{integer.  The number of principal components to use when producing
scree plot.  Only used if \code{scree = TRUE}.  Default is 10.}
}
\value{
The function returns a list containing the following: Lists called
  code{unweighted}, \code{spatial}, \code{temporal}, \code{spatiotemporal}
  are produced depending on what was included in \code{pca.wt}.  Each of
  these contains \code{scores} = principal components, \code{loads} =
  loadings, \code{var} = variance of each principal component,
  \code{var.prop} = proportion of total variance explained by each principal
  component, \code{cumvar} = cumulative proportion of total variance
  explained.

\code{pca.mode} contains the option used for \code{pca.mode}.

\code{pca.wt} contains the option used for \code{pca.wt}.

\code{call} contains the function call details.

\code{row.names} contains the row names of the data.frame on which PCA was
performed.  This will be row names of \code{x} if \code{pca.mode = "Smode"}
or column names of \code{x} if \code{pca.mode = "Tmode"}.

\code{col.names} contains the column names of the data.frame on which PCA was
performed.  This will be column names of \code{x} if \code{pca.mode =
"Smode"} or row names of \code{x} if \code{pca.mode = "Tmode"}.

\code{var.value} contains \code{var} from each of \code{unweighted},
\code{spatial}, \code{temporal} and \code{spatiotemporal} combined into a
single matrix.

\code{mean.timeseries} contains the average time series accross all sites.
This is stored for use when \code{plots="meanPM"} in \code{\link{plot_stpca}}
and is calculated as \code{apply(x, 1, mean)}.

A scree plot of the first \code{k} principal components is produced if
\code{scree = TRUE}, with the variance of each component on the y-axis.  The
text labels on the plot show the percentage variance explained by each
principal component.
}
\description{
PCA is performed on spatiotemporal data in T-mode or S-mode using singular
value decomposition and can be adjusted for spatial and/or temporal
autocorrelation if appropriate.  If spatial or temporal weights are used then
the returned loadings and principal components are backtransformed using
transformation described in paper (being written).
}
\details{
\code{x} must always have columns = monitoring sites and rows = time points,
regardless of the analysis the user intends to perform.  If \code{pca.mode =
"Tmode"} is specified, the function will automatically transpose the data.

\code{x} If column names in \code{spatial.wt} are not the same or in a
different order from \code{rownames(x)} then an error message is produced.

\code{scale} If TRUE then this is equivalent to performing PCA on the
correlation matrix.  If FALSE then this is equivalent to performing PCA on
the covariance matrix.

\code{pca.mode} PCA can be applied to spatiotemporal data in either S-mode or
T-mode.  In S-mode, the columns of \code{x} are monitoring sites while the
rows are time points.  In T-mode, the columns of \code{x} are time points
while the rows are monitoring sites.

\code{spatial.wt} The column names of the spatial weights matrix must be of
the same form and in the same order as \code{rownames(x)}.  An error is
produced if these do no match.  The user should order \code{x} and set
\code{rownames(x)} to correspond to the ordering in \code{spatial.wt}.
}
\examples{
library(stpca)

## get data.frame with no missing values
data(demoY)

## unweighted T-mode PCA
Tmode.pca.uw <- stpca(x=demoY, pca.mode="Tmode", pca.wt=c("unweighted"))

## spatial, temporal, and spatiotemporal weighted S-mode PCA

# create spatial weights
ssndata <- system.file("demoSSN/demoNet.ssn", package="stpca")
weightS <- createWeightS(ssndata=ssndata, afvcol="addfunccol")

# create temporal weights
weightT <- createWeightT(n=nrow(demoY), rho=0.5)

# check that the order of monitoring sites in x is the same as the
# columns in spatial.wt.  They are the same if the following results
# in 1.
mean(colnames(demoY) == colnames(weightS))

# weighted S-mode PCA
Smode.pca.all <- stpca(x = demoY, pca.mode = "Smode",
                       spatial.wt = weightS,
                       temporal.wt = weightT,
                       pca.wt = c("unweighted", "spatial",
                                  "temporal", "spatiotemporal"))

}

