% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stpca.R
\docType{class}
\name{StpcaModel-class}
\alias{StpcaModel-class}
\alias{StpcaModel}
\title{Structured PCA Model}
\description{
Structured PCA Model
}
\section{Methods}{

\describe{
\item{\code{update_beta(...)}}{Method docs go here}
}}

